requirements.txt
boto3==1.42.91
botocore==1.42.91
certifi==2026.2.25
charset-normalizer==3.4.7
Faker==40.15.0
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
kafka-python==2.3.1
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
